requirements/development.txt
# Base requirements
-r production.txt

# Development tools
ipython>=8.14.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
coverage>=7.3.2
pytest>=7.4.2
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
model-bakery>=1.12.0
freezegun>=1.3.2

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
django-stubs>=4.2.5
django-stubs-ext>=0.9.0

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
sphinx-autobuild>=2021.3.14

# Development server
django-browser-reload>=1.12.0
python-dotenv>=1.0.0
